bootimg: Fix dt.img generating issue (for OmniROM)
DT image variable is defined in CM build repo, but it's not in Omni one!
so the dt.img can't be generated nowhere!
and declaring this variable here ensures that it is defined before usage,
which makes compiling it possible on boot Omni and CM ROMs/based-ROMs.

Thanks to Humberto Borba for the hint.

Signed-off-by: Omar Einea <eineao@gmail.com>